+++ lib/src/discrete_mathematics/linear_algebra/exceptions/matrix_exception.dart
@@ -1,11 +1,7 @@
+import 'tensor_exception.dart';
+
 /// An exception that trows when performing improper matrix operations.
-class MatrixException implements Exception {
+class MatrixException extends TensorException implements Exception {
   /// Take [message] that describe reason of exception
-  MatrixException(this.message);
-
-  /// Field that contain reason of exception
-  String message;
-
-  @override
-  String toString() => '$message';
+  MatrixException(String message) : super(message);
 }