+++ lib/src/discrete_mathematics/linear_algebra/exceptions/tensor_exception.dart
@@ -1,0 +1,11 @@
+/// An exception that trows when performing improper tensor operations
+class TensorException implements Exception {
+  /// Take [message] that describe reason of exception
+  TensorException(this.message);
+
+  /// Field that contain reason of exception
+  String message;
+
+  @override
+  String toString() => '$message';
+}