+++ lib/src/discrete_mathematics/linear_algebra/exceptions/vector_exception.dart
@@ -1,11 +1,7 @@
-/// An exception that trows when performing improper vector operations.
-class VectorException implements Exception {
+import 'tensor_exception.dart';
+
+/// An exception that trows when performing improper vector operations
+class VectorException extends TensorException {
   /// Take [message] that describe reason of exception
-  VectorException(this.message);
-
-  /// Field that contain reason of exception
-  String message;
-
-  @override
-  String toString() => '$message';
+  VectorException(String message) : super(message);
 }