+++ lib/src/discrete_mathematics/linear_algebra/tensor/base/tensor_base.dart
@@ -1,13 +1,13 @@
 import '../../../../mixins/copyable_mixin.dart';
 
-/// Base class for tensor
-abstract class TensorBase<T> with CopyableMixin<TensorBase<T>> {
-  /// Creates instanse of TensorBase with dimension
-  TensorBase(this._dimension);
+/// Base class for tensors
+abstract class TensorBase with CopyableMixin<TensorBase> {
+  /// Creates instanse of TensorBase with dimensions
+  TensorBase(this._dimensions);
 
-  /// Holds dimension of this tensor
-  int _dimension;
+  /// Holds rank of this tensor
+  int _dimensions;
 
   /// Gets dimendion of this tensor
-  int get dimension => _dimension;
+  int get dimensions => _dimensions;
 }